+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+- Error out after a certain number of heartbeat errors
+
 ## [0.1.1] - 2018-11-05
 ### Fixed
 - Set browser display name based on `customLaunchers` key